TermStats as interface #1053
